requirements.txt
python-dotenv
requests
tqdm
hdbscan
umap-learn
pandas
numpy
